Use separate struct for send and receive
Split the RMAP header structs into send and receive versions in order to
remove the unused target/reply address field in the receive variants.
Previously this field was present but set to NULL with zero length, but
removing it completely cleans up the interface a bit.

Also update and unify the brief doxygen descriptions a bit, referring to
the headers as "Representation [...] for sending" or "Representation
[...] after reception and deserialization" instead of the previous less
clear "Deserialized representation [...]".