test: Set reply for "valid" write command
Set the reply command code when creating a valid write command which
includes a reply address, for testing Null pointer errors.

It seems like using a reply address without the reply command code is
valid according to the standard, but it is not obvious (the target must
not send a reply based on it, but the initiator may be allowed to set
it).