Add custom error for 32bit len too big for 24bit
Add a custom error for the given 32bit data length value being to big to
fit in the 24bit data field of the serialized packet.

This error is specific to the current library API, which accepts a 32bit
value for a 24bit entity

This error replaces the previously used RMAP_ECSS_TOO_MUCH_DATA error
which is now limited to describing too much data errors when
deserializing, thus following the definition in the ECSS RMAP standard
more by the letter.

This error is also deliberately separated from the (as of yet
unimplemented) read-modify-write data length error, since that error is
supposed to only cover read-modify-write commands, where any data
length except of 0, 2, 4, 6 and 8 is invalid.