Add more warnings
Add several warnings which did not seem to be included previously, but
seems relevant for this library.

Relevancy is determined based on the expected code features used by the
library, for example no extra floating point warnings are added since no
floating point features are expected to be used by the library.