Indicate that deserializing expects full packets
Reword the documentation for rmap_packet_serialize_inplace() to indicate
that the expected input data is a full RMAP packet rather than only an
RMAP header.

Taking the full packet as input is required since it is not possible to
separate the RMAP header without deserializing it first in order to
determine its type and size.