Verify data length & crc more in deserialize
Add verification of data length and data crc unconditionally for read
replies and conditionally based on the verify-before-write bit for write
commands when deserializing RMAP packets.

The length and crc check when verify-before-write is not set is
delegated to the caller, since some or all of the data should be written
without verification in this case.

Add two new status return values to indicate the new verification
failure cases.