Indicate output in serialization parameter doc.
Change the documentation for the serializing functions to indicate that
the destination parameter is an output parameter (instead of input, as
incorrectly indicated before).

This is based on a general rule for pointer parameters where if anything
is read via the pointer it is an input and if anything is written via
the pointer it is an output (not mutually exclusive).