Fix data length offset for verified write commands
Make sure to use the correct data length offset when verifying the size
and CRC for verified write commands.

Previously the same offset 8 was used for both read replies and write
commands, which is incorrect, since write commands have an offset of 12
instead.

This caused the length and CRC of verified write commands to be checked
using an invalid data length based on bytes from the address field.

This issue did not affect the extraction of the data length for
unverified write commands, which used the correct offset.

Add some tests which checks that when the instruction of ECSS test
pattern 0 is modified to contain an invalid packet type, deserializing
it reports appropriate errors, also check that all instructions
containing a valid write command combination are deserializing
correctly.

The test for valid write command combinations uncovered the now fixed
data length offset issue, since it tests verified write commands, such
commands not exists in the ECSS test patterns,, and was hence not
previously tested.