Include reply length in write command data offset
Make sure to include the reply address length in the data offset when
verifying write commands with verify-before-write set.

Previously the reply address length was not include which would lead to
false data verification failures for these packets.

Rework the calculation to get the offset based on the already calculated
header size, since this means the calculation is the same for both read
replies and write commands.

Add several tests which uses the existing ECSS test patterns where the
instruction field is modified (and the header CRC recalculated), in
order to test the behaviour when encountering:
* Reserved/invalid packet types.
* All valid command code combinations for this packet type.

Add a named variable for the padded reply address length stored in the
test pattern 2 command, since this length is different from the length
of the reply address present at the beginning of the test pattern 2
reply, where the padding is removed

Rework the parametrized test fixture for instruction modification and
split it into multiple fixtures named based on the packet type, in order
to only instantiate the tests with the test procedures which uses the
test patterns corresponding to the relevant packet type.

The instruction modifications tests needs to be specialized since
different packet types have different formats, and for example taking a
write command and setting its instruction field to indicate that it is a
read command would result in a completely broken packet.