Report error for reply without reply command code
Make sure to report an error when attempting to deserialize a packet
where the packet type indicates that it is a reply but where the command
codes indicates that no reply should have been sent for this command.

Previously no error was reported for such packet and it was deserialized
as if it was a valid reply, which it is not.

Add some additional modified instruction tests which checks this
behaviour, and update the previous tests which incorrectly assumed that
the previous behaviour was correct.