Deserialize whole instruction unconditionally
Deserialize and provide the contents of all instruction fields even if
packet type or the command codes are determined to be unused/invalid.

This provides no visible changes in behaviour for the time being.

This is done in order to provide support for returning a deserialized
header even if errors were detected in it. This is implicitly required
by the standard, since some of these errors should result in a reply,
and the header is needed to create the reply (in particular the reply
address and transaction id).

The packet type is now handled slightly differently, with the least
significant bit treated as the command/reply flag, and the most
significant bit treated as an independent reserved bit.

This allows unconditionally determining the packet type, which in the
future will allow unconditionally calculating the header CRC since the
header length is known based on the packet type and reply address
length. The standard seems to indicate that the CRC should indeed be
calculated unconditionally.