Deserialize headers even if invalid
Perform the deserializing of the header for commands and replies
unconditionally, provided there is enough data for the header based on
the packet type and reply address length (in the case of commands), and
accumulate any errors and report them at the end.

This allows providing a deserialized header for invalid commands, to
allow creating a reply as required by the standard for some invalid
header cases.

The deserialized header is also provided for replies, mostly to keep the
handling consistent. This deserializing is not needed since invalid
replies are directly discarded according to the standard.

Deserializing will now terminate early and not provide a deserialized
header only if:

* There is not enough data to fit any RMAP packet type header.
* There is not enough data to fit the RMAP packet type given by the
  instruction field (ignoring the reserved packet type bit).
* The header CRC is invalid.

Update the tests for unused packet types to avoid checking invalid
packet types that does not correspond to the test pattern used. This
would in most cases generate header CRC errors which could be tests, but
it is not an absolute certainty, so avoid the 1/256 chance that these
tests would behave unexpectedly due to hitting something which acts as a
valid CRC.

Add a test for invalid read command and read reply command codes, which
covers both reads without the reply bit set, and reads with the verify
before write bit set.