Allow serializing invalid headers
Rework serializing to allow serializing replies with an unused command
code, since this is required to send the reply with an error indication
based on a command with an unused command code.

Also allow serialising commands with an unused command code, to avoid
any special case and overloading of return values.

Split the unused packet type or command code error into 3 errors that
can be reported by the serializing and deserializing functions:

* Unused packet type.
* Unrepresentable command code value
* Unused command code

This allows the unused packet type error to be handled separately from
the unused command code error, allowing the caller to decide if the
target should send a reply (which is optional according to the RMAP
standard).

The unrepresentable command code value error is used when serializing to
indicate that the value given is not representable as an RMAP command
code in 4 bits.

The unused command code error is used to indicate that the header has
been serialized, but it contained errors.

Use RMAP_NO_REPLY to indicate that a reply header was missing the reply
command code flag, and that serializing has been aborted. Previously
this used the overloaded RMAP_ECSS_UNUSED_PACKET_TYPE_OR_COMMAND_CODE.

Update the parameters for the unrepresentable command codes to maintain
the representable bits set and add one extra bit above, instead of
adding one, which would cause all representable bits to be cleared via
carry.