Rework internals using accessors into raw header
Rework the internal handling of serialization and deserialization to use
accessor functions acting on the raw header representation.

This is done with the intention to eventually expose the "accessor
interface" as a user interface which does not use the custom object
representations for headers but instead keeps the header in the raw
format and operates on it via the accessors.

For now, rework the current user interface to use the new accessor
functions internally and maintain the same external behaviour.