Remove NULL reporting and just assert()
Remove the reporting of error code RMAP_NULLPTR for unexpected NULL
given by the library caller and instead use assert().

This pushes the burden of ensuring that no NULL pointers are passed in
onto the user, and also means that unless compiled with NDEBUG, this
library can terminate the whole application.

Reporting NULL errors would mean that all the raw packet buffer accessor
function would need to provide error reporting once exported, which
would mean that many of these functions would need to be reworked with
error reporting whose sole purpose would be to report NULL errors. This
is deemed undesirable and thus all NULL handling is reworked for the
sake of consistency.