Export rmap_{get,set}_*_data_length() & add tests
Rename get_header_data_length() and get_raw_data_length() with rmap_*
prefixes when exporting.

Since this is a 3-byte field, the previous byte accessor fixture cannot
be re-used.

The tests for the rmap_set_data_length() and
rmap_get_header_data_length() pair is implemented similarly as other
non-single-byte fields with a range of reply address size, limiting the
instantiations to the commands and replies which contains the data
length field.

For rmap_get_raw_data_length(), parametrization is limited to only the
reply address length, in order to avoid conditional logic to calculate
the header size in the test implementation (fully re-implementing
calculate_header_size() in the tests seems wrong).