test: Test reply address sizes even without reply
The RMAP standard indicates that the reply address length field and the
reply address field should always match, and is silent as to whether a
command without a reply should have a reply address length of 0.

It is assumed that the reply address length field and the reply address
field are completely independent of the with-reply flag.

Add testing of all reply address sizes even if the command is without
reply.