Split out verify_instruction()
In order to allow creating invalid commands and to allow sending correct
replies to invalid commands, the serialising functions need to allow
creating incorrect instruction fields.

Factor out the instruction verification to a separate function in order
to make it possible to allow creating incorrect instruction fields in
the future.