Rename reply_address0 -> reply_address
Using the reply_address0 naming in the exported function is unwanted
leaking of implementation details.

In order to also avoid using different parameter names between
declaration and definition, switch the usage so that the out pointer
parameter is the non-const pointer and the reply_address0 const pointer
is created locally inside the function scope instead.