Support creating invalid headers
Rework rmap_initialize_header() and make_instruction() to support
creating headers with invalid instruction fields, which is useful when
sending replies to invalid commands (which should include most of the
original instruction), and for testing purposes.

Re-add verification of the instruction field after initialising the header
in the rmap_serialize_* implementations to maintain the behaviour of
not allowing creation of invalid headers there.

Add new RMAP_PACKET_TYPE_COMMAND_RESERVED and
RMAP_PACKET_TYPE_REPLY_RESERVED constants in order to allow creating
invalid headers with reserved packet types.

Add a new RMAP_INVALID_PACKET_TYPE error code to indicate that the given
packet type cannot converted to a (valid or reserved) packet type
representation in an RMAP header.

Update the tests for rmap_initialize_header() to account for the new
behaviour where invalid headers are accepted.