test: Use generic patterns fixture to reduce tests
Convert the VerifyHeaderIntegrity fixture into a generic TestPatterns
fixture which can be used to check anything where the result is the same
for all test patterns and where knowning which test pattern is being
tested with is not important.

Use this to reduce the code duplication in existing tests for
rmap_calculate_and_set_header_crc() and
rmap_verify_header_instruction().