Remove rmap_get_raw_data_length()
This function was intended to be used for write without verify, but it
is not particularly useful in that case, since the assumption that the
last byte of the packet is the CRC and should not be written does not
necessarily hold.

If not verifying before writ and the packet is smaller than indicated by
the header data length, then all data after the header should be
written, including the last byte of the packet (and then an early EOP
error should be reported).

Hence it is more useful to simply use the difference between the packet
size and rmap_calculate_header_size() to determine the amount of data to
write before verification.

Remove rmap_get_raw_data_length() and associated tests.