make: Avoid re-archiving each time
When using the archive(member) in make to refer to a member inside an
archive as targets or prerequisites, the timestamps of the members in
the archive is used to determine if they need to be re-added to the
archive.

Many distributions currently compile the ar tool with deterministic mode
enabled by default which means that member timestamps are always set to
zero. This results in the member inside the archive always being older
than the member outside the archive according to the timestamps, which
results in make always trying to re-add the member to the archive.

Avoid this unconditional re-archiving by removing the use of make
archive members completely and using a non-implicit rule to create the
archive directly from the member files, without the intermediate member
in archive target.

The minor downsides of avoiding archive members are:
* Implicit rules cannot be used.
* If there are multiple members in an archive, it is no longer possible
  to only update one of them (this would only really matter if the
  archive was huge).

It would be possible to force the use of timestamps by adding the U flag
to the ar invocation. But this would make the archive non-deterministic,
and determinism is desirable.