Use protocol format for pkt type & command codes
Use a representation for packet type and command code enum constants
which matches the protocol representation in RMAP header fields.

Use the matching representation to simplify the implementation.

Update a test which relies on the packet type numeric values to get the
first non-representable value.

The reason for previously using a representation which did not match the
RMAP protocol representation is an argument that this is a
library-specific representation, and thus should not be relied upon to
match the protocol format, but the convenience of having them match
seems like a better trade-off.