Remove command code shift & mask flag defines
Since the RMAP_COMMAND_CODE_* enum constants now correspond to the
protocol format, they can be used when masking the command code field,
which makes the mask and shift defines for each individual command code
flag unnecessary, so remove them.

Update packet type accessor functions to avoid duble masking and to use
the RMAP_PACKET_TYPE_COMMAND as a mask since it now corresponds to the
protocol format.

When checking the reserved packet type bit, an explicit 0x2 is still
used since otherwise the packet type constants would need to be defined
as flags instead of distinct types, which would make the use a bit less
intuitive: If passing the packet type as OR:ed flags (for example
COMMAND | RESERVED) the input parameter would need to be a plain int,
and thus provide less compile-time checking.

Keep the mask defines for the whole packet type, command code, and reply
address length fields, but change 3 -> 0x3 to hint extra clearly that
they are masks.

Update tests to use enum constants and field shifts instead of literal 1
and flag-specific shifts for both command codes and packet types.