Use structs for internal field access
Use structs as internal helpers to set and get fields from the header,
in order to remove most of the hardcoded field offsets from the code.

These structs only contain octet fields in order to ensure that no
padding is present and to make it obvious that the multibyte fields are
raw representations which need conversion.

For commands, the reply address field is of variable size, hence "first
static part" and "second static part" structs are used for the static
parts of the header. Also add helper functions for getting the location
of the second static part based on the reply address size.