test: Use struct for test pattern + metadata
Use a common struct as a container for test patterns, allowing header
offset (due to prefix addresses) and reply address length to be
associated with the test pattern in a generic way.

Also add some common collections of test pattern for use in
parameterized tests.

Rework tests to access the test patterns using the new format.

Rework RmapInitializeHeaderPatternsShouldNotChange test to avoid
illegally using rmap_get_reply_address on replies and instead use the
test pattern metadata to get the reply address length.

Some tests which previously used explicit test patterns are now easier
to parameterize, so this is done. This includes CRC tests which are
extended to test all patterns.