test: Remove test with cast undefined behaviour
Converting to an enumeration type from a value that is outside the range
of the defined enumeration values is undefined behaviour in C++.

Remove test instantiations that use such casts in order to test invalid
packet types.

This is not undefined behaviour in C, so testing it would make sense,
but doing so would involve quite a bit of complexity; it is skipped for
the time being.