Add clang-format config and reformat all code
The configuration is chosen to match the current code style fairly
close.

Style options are chosen to prioritize formatting that is easy and
reasonably natural to write manually, instead of formatting which only
would be convenient to maintain using a formatting tool.

In many cases, clang-format produces worse formatting than before, but
this is accepted in order to prioritize consistency and to facilitate
automated checks. (In particular initializer lists and the
misidentification of '<<' as stream operators are quite ugly.)