Add trailing commas for prettier clang-format
Add trailing comma separators at the end of multi-line initializer
lists in order for clang-format to change

  int a[] = {
    (...)
    x};

to instead

  int a[] = {
    (...)
    x,
  };

which is subjectively considered nicer. In particular since it avoids
a special case for the last element.

The trailing comma is omitted for single-line initializer lists in order
to avoid clang-format splitting them into multiple lines.