node: Return errors instead of "collecting"
Return error information via rmap_node_handle_incoming() instead of
storing it in a member of the node context struct.

The storing was done in order to match the wording of the RMAP standard
regarding the node collecting error information, but it likely makes
more sense to provide this to the caller directly and leave the
responsibility of if and how to collect this information to the user of
the library instead.

Also merge the node statuses back into the base rmap_status enum type,
in order to reduce the duplication. This does however mean that rmap.h
is to some degree node-aware, even if compiled without any node support.