node: Note callbacks must use valid retvals
Add a strong note that request callbacks must be implemented by the user
to only return one of the return values applicable to the specific
callback.

It is treated as undefined behaviour if this requirement is not met. In
practice, the current implementation will either:
* Act as if the callback returned RMAP_STATUS_FIELD_CODE_SUCCESS.
* Trigger an assert() if asserts are enabled.

Providing error reporting for incorrectly implemented callbacks was
considered, but in some cases this would mean either:
* Sending an RMAP error reply using some generic status code, likely
  RMAP_STATUS_FIELD_CODE_GENERAL_ERROR_CODE.
* Leaking the reply allocation.
None of these alternatives seem particularly nice, so it is skipped
altogether.