rmap: "unrepresentable" -> "cannot be represented"
Reword "unrepresentable packet type" to "a value which cannot be
represented as a packet type in an RMAP packet" (and similar for command
code).

This is intended to better convey that the errors indicate that the
given value has has too many bits to fit in its field and/or is a
negative value, so it cannot even be used to form an invalid RMAP
header.