node: Write header after data for read & RMW
This is intended to allow the user of the library to use the header
space in the allocation as a scratch space when performing a read.

This can for example be useful if reading partial data which needs to be
obtained as a whole and then adjusted to match the actual request.