github: Correct runner image parameterization
Using multiple strings for "runs-on:" specifies an AND condition, so
with

  runs-on: [ubuntu-latest, ubuntu-24.04]

only a runner which matches both will be used, which at the time of
writing with just stall the action indefinitely.

Convert the parameterization to use strategy.matrix to make sure the
action runs on both (at time of writing, ubuntu-latest is still
ubuntu-22.04).