node: Avoid list end issue with older doxygen
Doxygen versions 1.8.19 up to and including 1.9.1 has an issue where
list end termination via a period does not work properly
(https://github.com/doxygen/doxygen issue 8163). Ubuntu 22.04 has
version 1.9.1 of doxygen.

Restructure the documentation with a nested list to work around this
issue.

The preliminary plan is to revert this workaround once the older
versions of doxygen are no longer in common use.

Using conditional documentation sections for different doxygen versions
was explored briefly, but was abandoned due to the inconvenience of
injecting conditionals and the minor overall benefit.