node: Allow adjusting read data destination
If more data than requested needs to be read in order to then adjust the
start of the data forwards to the correct offset,

Due to the node library writing the RMAP reply address and header to the
start of the destination returned by the allocation callback, it is not
possible to adjust the read data in any way.

Adjusting the data destination as part of the read or RMW request
callback may be desirable in the following cases:

* Reading must be done in fixed size blocks, the desired start of
  data is at an offset into a block, and avoiding additional copying is
  important (and the allocation is large enough to still fit the data if
  using an offset).

* Reading must be done into a specific destination that is different
  from the one given by the allocation callback.

In order to support this, add a pointer indirection to the output read
data parameter, allowing the read and RMW request callbacks to change
its destination.

The node handling needs to build the RMAP reply around the data, so the
new destination must still have enough space for the reply address and
header before, and the data CRC after.

Adapt tests and examples to this new interface, and add some dedicated
tests for this functionality.